Make build work with shell.
This is the first build that actually works.
Part of the problem before was that the extration for the DevKit was not
not being performed properly.  This commit fixes that issue alogn with
some others.

We are almost finished.  We just need to make sure that ConEmu link
wokrs and runs .bashrc/set-env.sh automatically.